// src/app/api/models/quotation-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { QuotationLineRequest } from '../models/quotation-line-request';

/**
 * Lines are sent whole: on update, a `lines` list replaces what was there.
 * Leave `lines` out to change the header only.
 */
export interface QuotationRequest {
  customer: number;
  lines?: Array<QuotationLineRequest>;
  note?: string;
  quote_date?: string;
  terms?: string;
  valid_until?: string;
}